refactor(ai): functional provider config layer + cache-warm dedup
Replace the ProviderConfigurationService (2285 lines) and ProviderManager
(459 lines) god-classes with functional modules under ai/providers/config/
(free functions + explicit AiProviderCtx, no BaseModel), keeping the
two-altitude compute/cache split: compute* (DB-direct, feeds the org cache)
vs get* (cache-backed public API).

- Fix switchProviderType cache invalidation: the action now fires
  ai-provider.type-switched after the write (the bare mutation only wrote
  ProviderPreference, leaving stale config for the 15-min TTL).
- Repoint all consumers (tRPC router, both orchestrators, embedding-service,
  cache compute providers, debug script) and delete the classes.

Cache-warm performance:
- resolveCredentials accepts a precomputed ProviderCredentialContext so the
  aiCredentials compute loop resolves each provider's config + BYO default
  key once instead of re-assembling per registry model (was ~N full
  single-provider assemblies per warm).
- OrganizationAiQuota is fetched once per compute and threaded into the
  per-provider assembly instead of one identical query per provider.
