+++ packages/lib/src/ai/providers/index.ts
@@ -1,5 +1,29 @@
 // packages/lib/src/ai/providers/index.ts
 
+// Functional provider config layer (replaces ProviderConfigurationService + ProviderManager)
+export type { AiProviderCtx } from './config'
+export {
+  computeProviderConfig,
+  computeProviderConfigs,
+  deleteCustomModel,
+  deleteProvider,
+  getCredentials,
+  getEffectiveConfig,
+  getModelTypeForModel,
+  getProviderConfig,
+  getProviderConfigs,
+  getSystemCredentials,
+  getUnifiedModelData,
+  isModelCompatible,
+  removeCustomCredentials,
+  resolveCredentials,
+  saveCustomModel,
+  saveProvider,
+  switchProviderType,
+  testProvider,
+  toggleModel,
+  updateModelConfig,
+} from './config'
 // Connection blueprint mapping
 export {
   AI_PROVIDER_CONNECTION_KEY,
@@ -7,9 +31,6 @@ export {
   aiProviderConnectionKey,
   CONNECTION_KEY_AI_PROVIDER,
 } from './connection-provider-map'
-// Main service classes
-export { ProviderConfigurationService } from './provider-configuration-service'
-export { ProviderManager } from './provider-manager'
 export { ProviderRegistry } from './provider-registry'
 export { type SystemModelDefaultEntity, SystemModelService } from './system-model-service'
 // Re-export commonly used types for convenience
