+++ packages/lib/src/cache/providers/ai-credentials-provider.ts
@@ -10,26 +10,37 @@ import type { CacheProvider } from '../org-cache-provider'
  */
 export const aiCredentialsProvider: CacheProvider<Record<string, CredentialsResponse>> = {
   async compute(orgId, db) {
-    const { ProviderConfigurationService } = await import(
-      '../../ai/providers/provider-configuration-service'
+    const { computeProviderConfigs } = await import('../../ai/providers/config/assemble')
+    const { loadProviderCredentialContext, resolveCredentials } = await import(
+      '../../ai/providers/config/runtime-credentials'
     )
-    const service = new ProviderConfigurationService(db, orgId, 'system')
-    const { configurations } = await service.getConfigurations()
+    const ctx = { db, organizationId: orgId, userId: 'system' }
+    const { configurations } = await computeProviderConfigs(ctx)
     const credentials: Record<string, CredentialsResponse> = {}
 
     for (const [provider, config] of Object.entries(configurations)) {
+      // Resolve the per-provider invariants (config + BYO default key) once, reusing the
+      // already-assembled `config` — every model below threads it instead of re-assembling.
+      const providerCtx = await loadProviderCredentialContext(ctx, provider, config)
+
       // Cache provider-level credentials (no model/modelType)
-      const providerCreds = await service.getCurrentCredentials(provider, null, null)
-      credentials[`${provider}:__provider__`] = providerCreds
+      credentials[`${provider}:__provider__`] = await resolveCredentials(
+        ctx,
+        provider,
+        null,
+        null,
+        providerCtx
+      )
 
       // Cache per-model credentials for each configured model
       for (const model of config.models) {
-        const modelCreds = await service.getCurrentCredentials(
+        credentials[`${provider}:${model.modelId}:${model.modelType}`] = await resolveCredentials(
+          ctx,
           provider,
           model.modelId,
-          model.modelType as ModelType
+          model.modelType as ModelType,
+          providerCtx
         )
-        credentials[`${provider}:${model.modelId}:${model.modelType}`] = modelCreds
       }
     }
